+++ app/config/config.yml
@@ -46,6 +46,7 @@ twig:
 # Doctrine Configuration
 doctrine:
     dbal:
+        server_version: 5.7
         driver:   pdo_mysql
         host:     "%database_host%"
         port:     "%database_port%"
