+++ src/AppBundle/Controller/ArticleController.php
@@ -4,8 +4,8 @@ namespace AppBundle\Controller;
 
 use AppBundle\Entity\Article;
 use AppBundle\Entity\Comment;
-use AppBundle\Form\Article\ArticleType;
-use AppBundle\Form\Comment\CommentType;
+use AppBundle\Form\ArticleType;
+use AppBundle\Form\CommentType;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
