+++ src/AppBundle/Form/CommentType.php
@@ -1,22 +1,19 @@
 <?php
 
-namespace AppBundle\Form\Article;
+namespace AppBundle\Form;
 
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-
-class ArticleType extends AbstractType
+class CommentType extends AbstractType
 {
     /**
      * {@inheritdoc}
      */
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
-        $builder->add('date', DatePickerType::class)
-                ->add('title')
-                ->add('content', TinyMCEType::class);;
+        $builder->add('content');
     }
 
     /**
@@ -26,7 +23,7 @@ class ArticleType extends AbstractType
     {
         $resolver->setDefaults(
             array(
-                'data_class' => 'AppBundle\Entity\Article',
+                'data_class' => 'AppBundle\Entity\Comment',
             )
         );
     }
@@ -36,7 +33,7 @@ class ArticleType extends AbstractType
      */
     public function getBlockPrefix()
     {
-        return 'appbundle_article';
+        return 'appbundle_comment';
     }
